rework the Netsim Internals to support handling pointers for FFI instead of relying only on byte array.
This effectively allows the users to call the C interface with
bytes (in the message structure they can add the pointer to an array
and the appropriate size). Or they can put in the pointer to any
object and the "simulated size".

If the package end up being dropped the `on_drop` function will
be called, allowing the user to `free` the allocated assets.