Implement real congestion simulation
In order to do that we are utilising a different approach to computing
when a message is meant to be consummed by a receiver:

1. first we apply the latency, this is the minimum that is always
   going to be applied to any message, it is computing at the reception
   of the message
2. every tick we count how much the upload on the receiver and link
   and then on the download can be computed.

The latest point is what is going to manage the congestion issue:

we compute 3 counters: one for how much data has been used by the
sender's upload, how much has been used by the link and how much
has been used by the receiver's download.

If there is enough bandwidth remaining for part or all the message
to be uploaded by the sender we consum this bytes. If enough is
remaining on the link we consum some or all of the uploaded data
and if enough is remaining on the receiver's download we consum
all or part of the data in the link in the download.

Once we reach in the receiving end the total of the message it can
then be removed from the queues of messages transiting through the
simulator and can then be put on the receiver's `SimSocket` queue.